+++ public/js/master/distributionList.js
@@ -79,9 +79,11 @@ $(function () {
 
     function edit(target) {
         var id = $('td[name="id"]', $(target).closest('tr')).html();
+        var codeValue = $('td[name="codeValue"]', $(target).closest('tr')).html();
         var name = $('td[name="name"]', $(target).closest('tr')).html();
         var modal = $('#editModal');
         modal.find('input[name=id]').val(id);
+        modal.find('input[name=codeValue]').val(codeValue);
         modal.find('input[name=name]').val(name);
         modal.modal();
     }