+++ dst/controllers/event/screeningEvent.js
@@ -339,14 +339,14 @@ function createEventFromBody(req) {
             throw new Error('上映スクリーン名が見つかりません');
         }
         const ticketTypeGroup = yield offerService.findTicketTypeGroupById({ id: body.ticketTypeGroup });
-        const { data } = yield serviceTypeService.search({
-            project: { ids: [req.project.id] },
-            ids: [ticketTypeGroup.itemOffered.serviceType.id]
+        const searchServiceTypesResult = yield serviceTypeService.search({
+            limit: 1,
+            codeValue: { $eq: ticketTypeGroup.itemOffered.serviceType.codeValue }
         });
-        if (data.length === 0) {
-            throw new Error('興行区分が見つかりません');
+        const serviceType = searchServiceTypesResult.data.shift();
+        if (serviceType === undefined) {
+            throw new Error('興行タイプが見つかりません');
         }
-        const serviceType = data[0];
         let offersValidAfterStart;
         if (body.endSaleTimeAfterScreening !== undefined && body.endSaleTimeAfterScreening !== '') {
             offersValidAfterStart = Number(body.endSaleTimeAfterScreening);
@@ -485,7 +485,7 @@ function createMultipleEventFromBody(req, user) {
         const ticketTypeGroups = searchTicketTypeGroupsResult.data;
         const searchBoxOfficeTypeGroupsResult = yield serviceTypeService.search({
             limit: 100,
-            project: { ids: [req.project.id] }
+            project: { id: { $eq: req.project.id } }
         });
         const serviceTypes = searchBoxOfficeTypeGroupsResult.data;
         const attributes = [];
@@ -519,7 +519,7 @@ function createMultipleEventFromBody(req, user) {
                     if (ticketTypeGroup === undefined) {
                         throw new Error('Ticket Type Group');
                     }
-                    const serviceType = serviceTypes.find((t) => t.id === ticketTypeGroup.itemOffered.serviceType.id);
+                    const serviceType = serviceTypes.find((t) => t.codeValue === ticketTypeGroup.itemOffered.serviceType.codeValue);
                     if (serviceType === undefined) {
                         throw new Error('Service Type Not Found');
                     }