+++ src/controllers/event/screeningEvent.ts
@@ -334,14 +334,15 @@ async function createEventFromBody(req: Request): Promise<chevre.factory.event.s
     }
 
     const ticketTypeGroup = await offerService.findTicketTypeGroupById({ id: body.ticketTypeGroup });
-    const { data } = await serviceTypeService.search({
-        project: { ids: [req.project.id] },
-        ids: [ticketTypeGroup.itemOffered.serviceType.id]
+
+    const searchServiceTypesResult = await serviceTypeService.search({
+        limit: 1,
+        codeValue: { $eq: ticketTypeGroup.itemOffered.serviceType.codeValue }
     });
-    if (data.length === 0) {
-        throw new Error('興行区分が見つかりません');
+    const serviceType = searchServiceTypesResult.data.shift();
+    if (serviceType === undefined) {
+        throw new Error('興行タイプが見つかりません');
     }
-    const serviceType = data[0];
 
     let offersValidAfterStart: number;
     if (body.endSaleTimeAfterScreening !== undefined && body.endSaleTimeAfterScreening !== '') {
@@ -489,7 +490,7 @@ async function createMultipleEventFromBody(req: Request, user: User): Promise<ch
 
     const searchBoxOfficeTypeGroupsResult = await serviceTypeService.search({
         limit: 100,
-        project: { ids: [req.project.id] }
+        project: { id: { $eq: req.project.id } }
     });
     const serviceTypes = searchBoxOfficeTypeGroupsResult.data;
 
@@ -525,7 +526,7 @@ async function createMultipleEventFromBody(req: Request, user: User): Promise<ch
                 if (ticketTypeGroup === undefined) {
                     throw new Error('Ticket Type Group');
                 }
-                const serviceType = serviceTypes.find((t) => t.id === ticketTypeGroup.itemOffered.serviceType.id);
+                const serviceType = serviceTypes.find((t) => t.codeValue === ticketTypeGroup.itemOffered.serviceType.codeValue);
                 if (serviceType === undefined) {
                     throw new Error('Service Type Not Found');
                 }