+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 興行区分をカテゴリーコードに統合
+- 配給区分をカテゴリーコードに統合
 
 ### Deprecated