+++ CHANGELOG.md
@@ -18,6 +18,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v1.6.0 - 2020-02-03
+
+### Changed
+
+- 興行区分をカテゴリーコードに統合
+- 配給区分をカテゴリーコードに統合
+- update @chevre/api-nodejs-client
+
 ## v1.5.0 - 2020-01-29
 
 ### Changed